+++ src/ui/allScreens.js
@@ -25,7 +25,7 @@ import NetworkSearchScreen from './screens/networksearch';
 import HomeSearchScreen from './screens/homesearch';
 import AddItemScreen from './screens/additem';
 import TestScreen from './screens/test';
-import UserSheetScreen from './screens/userSheet';
+import UserSheet from './screens/userSheet';
 import MyAvatar from './components/MyAvatar';
 import {InteractionScreen} from './screens/interactions';
 import AddLineupSheet from './screens/addlineupsheet';
@@ -41,7 +41,6 @@ import React from 'react';
 import {View} from 'react-native';
 import UnsaveScreen from "./screens/unsave";
 import ChangeDescriptionScreen from "./screens/changeDescription";
-import FollowUser from "./screens/followUser";
 import ChangeLineupName from "./screens/changeLineupName";
 import MoveInScreen from "./screens/moveinscreen";
 import InAppNotif from "./screens/inAppNotif";
@@ -107,12 +106,11 @@ export default function registerScreens(store, Provider) {
     wrap('goodsh.MyAvatar', () => MyAvatar, store, Provider);
     wrap('goodsh.UserNav', () => UserNav, store, Provider);
     wrap('goodsh.LineupNav', () => LineupNav, store, Provider);
-    wrap('goodsh.UserSheetScreen', () => UserSheetScreen, store, Provider);
+    wrap('goodsh.UserSheet', () => UserSheet, store, Provider);
     wrap('goodsh.AddLineupSheet', () => AddLineupSheet, store, Provider);
     wrap('goodsh.UnsaveScreen', () => UnsaveScreen, store, Provider);
     wrap('goodsh.ChangeDescriptionScreen', () => ChangeDescriptionScreen, store, Provider);
     wrap('goodsh.ChangeLineupName', () => ChangeLineupName, store, Provider);
-    wrap('goodsh.FollowUser', () => FollowUser, store, Provider);
     wrap('goodsh.PlacesAutocomplete', () => PlacesAutocomplete, store, Provider);
     wrap('goodsh.MoveInScreen', () => MoveInScreen, store, Provider);
     wrap('goodsh.InAppNotif', () => InAppNotif, store, Provider);
